fix(mutexbot): avoid race condition when fetching reservation details
When trying to reserve a resource fails, the action tries to fetch
details for this reservation. When the previous reservation expires
between trying to reserve and fetching the reservation details, the
action exited with an error because it couldn't read reservation details
from the list. This change here makes the logging conditional on
actually finding an active reservation, instead of failing if it can't
find a reservation.