Introduce reserve-exclusive to wait for a resource.
Mutexbot API semantics allow the `reserve` API call to silently update an
existing reservation. This makes it hard to use `mutexbot` as a concurrency
control mechanism in GitHub workflows that share the same API key.

This PR introduces a new mode for mutexbot, called `reserve-exclusive`. This
new mode checks for the existence of active reservations before trying to
reserve a resource. If an active reservation exists, the code keeps checking
periodically, until it is finally able to get a reservation.

Example output:

```shell
$ ./mutexbot reserve-exclusive test-resource "" "1m"
[2025-09-17T13:54:36Z INFO  mutexbot] Resource reserved successfully

$ ./mutexbot reserve-exclusive test-resource "" "1m"
[2025-09-17T13:54:39Z INFO  mutexbot] Existing reservation by user testBot with reason "". Expires at: 2025-09-17 13:55:36.455 UTC.
[2025-09-17T13:54:39Z INFO  mutexbot] Resource is reserved, waiting 57.0 seconds before retrying...
[2025-09-17T13:55:37Z INFO  mutexbot] Resource reserved successfully
```