fix(deploy-queue): allow running with missing migrations
As the deploy-queue runs in a bunch of different places and in some
cases with git repos that are not up-to-date, we'll regularly have older
versions of deploy-queue still in service. Requiring all migrations to
be present in the source code means that as soon as a new migration is
deployed, all older versions will break. That is unacceptable, so we'll
have to allow running with missing migrations.